requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
boto3==1.34.34
python-dotenv==1.0.0
marshmallow==3.20.1
Flask-CORS==4.0.0
gunicorn==21.2.0
Pillow==10.2.0
requests==2.31.0
google-genai>=0.1.0
